+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 CHANGELOG
 =========
 
-unreleased
-----------
+1.9.1
+-----
 * The CHANGELOG is now part of the packages.
 * Improved unicode handling in Python2 (#78)
 * Add `tsv` output format for **shodan host** (#65)